Added Group Messages and DMs into one Object

+++ Backend/Application/src/main/java/CampusConnect/Database/Models/Messages/DisplayChat.java
@@ -2,7 +2,7 @@ package CampusConnect.Database.Models.Messages;
 
 public class DisplayChat
 {
-    public int Id; // Either User ID or Session Id
+    public Long Id; // Either User ID or Session Id
     public String Name; // Either Session Name itself or User Name
     public boolean isGroupChat;
